+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/worldgen/processors/BeeDungeonProcessor.java
@@ -8,11 +8,15 @@ import com.telepathicgrunt.the_bumblezone.modcompat.ModCompat;
 import com.telepathicgrunt.the_bumblezone.modinit.BzBlocks;
 import com.telepathicgrunt.the_bumblezone.modinit.BzFluids;
 import com.telepathicgrunt.the_bumblezone.modinit.BzProcessors;
+import com.telepathicgrunt.the_bumblezone.modinit.BzTags;
 import com.telepathicgrunt.the_bumblezone.utils.GeneralUtils;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.HolderSet;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.CandleBlock;
 import net.minecraft.world.level.block.state.BlockState;
@@ -22,6 +26,8 @@ import net.minecraft.world.level.levelgen.structure.templatesystem.StructurePlac
 import net.minecraft.world.level.levelgen.structure.templatesystem.StructureProcessor;
 import net.minecraft.world.level.levelgen.structure.templatesystem.StructureProcessorType;
 import net.minecraft.world.level.levelgen.structure.templatesystem.StructureTemplate;
+
+import java.util.Optional;
 
 /**
  * POOL ENTRY MUST BE USING legacy_single_pool_element OR ELSE THE STRUCTURE BLOCK IS REMOVED BEFORE THIS PROCESSOR RUNS.
@@ -67,9 +73,12 @@ public class BeeDungeonProcessor extends StructureProcessor {
                             blockState = BzBlocks.HONEY_CRYSTAL.get().defaultBlockState();
                         }
                         else if (random.nextFloat() < 0.6f) {
-                            blockState = GeneralUtils.VANILLA_CANDLES.get(random.nextInt(GeneralUtils.VANILLA_CANDLES.size()));
-                            blockState = blockState.setValue(CandleBlock.CANDLES, random.nextInt(4) + 1);
-                            blockState = blockState.setValue(CandleBlock.LIT, true);
+                            Optional<HolderSet.Named<Block>> optionalBlocks = BuiltInRegistries.BLOCK.getTag(BzTags.BEE_DUNGEON_POSSIBLE_CANDLES);
+                            if (optionalBlocks.isPresent()) {
+                                blockState = optionalBlocks.get().get(random.nextInt(optionalBlocks.get().size())).value().defaultBlockState();
+                                blockState = blockState.setValue(CandleBlock.CANDLES, random.nextInt(4) + 1);
+                                blockState = blockState.setValue(CandleBlock.LIT, true);
+                            }
                         }
                         else {
                             blockState = Blocks.CAVE_AIR.defaultBlockState();
@@ -80,9 +89,12 @@ public class BeeDungeonProcessor extends StructureProcessor {
                             blockState = BzBlocks.HONEY_CRYSTAL.get().defaultBlockState();
                         }
                         else if (random.nextFloat() < 0.35f) {
-                            blockState = GeneralUtils.VANILLA_CANDLES.get(random.nextInt(GeneralUtils.VANILLA_CANDLES.size()));
-                            blockState = blockState.setValue(CandleBlock.CANDLES, random.nextInt(random.nextInt(4) + 1) + 1);
-                            blockState = blockState.setValue(CandleBlock.LIT, true);
+                            Optional<HolderSet.Named<Block>> optionalBlocks = BuiltInRegistries.BLOCK.getTag(BzTags.BEE_DUNGEON_POSSIBLE_CANDLES);
+                            if (optionalBlocks.isPresent()) {
+                                blockState = optionalBlocks.get().get(random.nextInt(optionalBlocks.get().size())).value().defaultBlockState();
+                                blockState = blockState.setValue(CandleBlock.CANDLES, random.nextInt(random.nextInt(4) + 1) + 1);
+                                blockState = blockState.setValue(CandleBlock.LIT, true);
+                            }
                         }
                         else {
                             blockState = Blocks.CAVE_AIR.defaultBlockState();
@@ -93,9 +105,12 @@ public class BeeDungeonProcessor extends StructureProcessor {
                             blockState = BzBlocks.HONEY_CRYSTAL.get().defaultBlockState();
                         }
                         else if (random.nextFloat() < 0.2f) {
-                            blockState = GeneralUtils.VANILLA_CANDLES.get(random.nextInt(GeneralUtils.VANILLA_CANDLES.size()));
-                            blockState = blockState.setValue(CandleBlock.CANDLES, random.nextInt(random.nextInt(4) + 1) + 1);
-                            blockState = blockState.setValue(CandleBlock.LIT, true);
+                            Optional<HolderSet.Named<Block>> optionalBlocks = BuiltInRegistries.BLOCK.getTag(BzTags.BEE_DUNGEON_POSSIBLE_CANDLES);
+                            if (optionalBlocks.isPresent()) {
+                                blockState = optionalBlocks.get().get(random.nextInt(optionalBlocks.get().size())).value().defaultBlockState();
+                                blockState = blockState.setValue(CandleBlock.CANDLES, random.nextInt(random.nextInt(4) + 1) + 1);
+                                blockState = blockState.setValue(CandleBlock.LIT, true);
+                            }
                         }
                         else {
                             blockState = Blocks.CAVE_AIR.defaultBlockState();